% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMISE.R
\name{AMISE}
\alias{AMISE}
\title{AMISE bandwidth}
\usage{
AMISE(x, y = NULL)
}
\arguments{
\item{x}{the number of variables (if y is given), or a data frame or a matrix (if y is missing).}

\item{y}{the number of observations. If y is missing then x is interpreted as the data matrix.}
}
\value{
AMISE bandwidth.
}
\description{
Calculate the AMISE bandwidth from either a data frame, or from the number of observations and the dimension of the data.
}
\details{
IMPORTANT NOTE: The standard deviation of each variable is omitted in this formula.
}

