% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust10.R
\docType{data}
\name{clust10}
\alias{clust10}
\title{1000 5-dimensional data points that form ten clusters}
\format{data frame}
\description{
Generated from the genRandomClust() function of the "clusterGeneration" package with separation value 0.2.
}
\keyword{datasets}

