% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust3.R
\docType{data}
\name{clust3}
\alias{clust3}
\title{90 2-dimensional data points that form three clusters}
\format{data frame}
\description{
Randomly generated from three normal distributions.
}
\keyword{datasets}

