% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust5.R
\docType{data}
\name{clust5}
\alias{clust5}
\title{500 5-dim points scattered in 5 clusters}
\format{data frame}
\description{
500 5-dim points in 5 clusters. Generated from the genRandomClust() function of the "clusterGeneration" package with separation value 0.1.
}
\keyword{datasets}

