% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dat_gene.R
\docType{data}
\name{dat_gene}
\alias{dat_gene}
\title{gene expression data of 38 patients (243 genes)}
\format{matrix}
\description{
38 by 243 matrix. Each row is a patient. Each column is a gene.
}
\keyword{datasets}

