% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCentersAutoD.R
\name{FindCentersAutoD}
\alias{FindCentersAutoD}
\title{Automatically find centers with diagonal f(x) vs delta(x) thresholds}
\usage{
FindCentersAutoD(f, delta, nclust, rm.dup = TRUE)
}
\arguments{
\item{f}{vector of local distance f(x)}

\item{delta}{vector of minimal distances to higher ground delta(x)}

\item{nclust}{number of clusters. Can be a vector.}

\item{rm.dup}{boolean. If TRUE (default) remove duplicated center vectors.}
}
\value{
a list of vectors. Each vector gives the locations of centers.
}
\description{
Automatically find centers with diagonal f(x) vs delta(x) thresholds
}
\author{
Ethan Xu
}

