% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCentersAutoV.R
\name{FindCentersAutoV}
\alias{FindCentersAutoV}
\title{Automatically find centers with vertical f(x) thresholds}
\usage{
FindCentersAutoV(f, delta, f.cut = c(0.1, 0.2, 0.3), nclust, rm.dup = TRUE)
}
\arguments{
\item{f}{vector of local distance f(x)}

\item{delta}{vector of minimal distances to higher ground delta(x)}

\item{f.cut}{number between (0, 1) or numeric vector of numbers between (0,1). f.cut is used in centroids = "auto" to automatically select cluster centroids from the decision plot. Points with f(x) > f.cut and high delta(x) are selected as one set of candidate centroids (see details). Default = c(0.1, 0.2, 0.3).}

\item{nclust}{number of clusters. Can be a vector.}

\item{rm.dup}{boolean. If TRUE (default) remove duplicated center vectors.}
}
\value{
a list of vectors. Each vector gives the locations of centers.
}
\description{
Automatically find centers with vertical f(x) thresholds
}
\author{
Ethan Xu
}

