% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCluster.R
\name{findCluster}
\alias{findCluster}
\title{User-interactive Finding Clusters}
\usage{
findCluster(x, mycols = NULL)
}
\arguments{
\item{x}{a list. Return value from findRd2(). The list contains "f", "delta",
"dat", and "distm".}

\item{mycols}{a vector of character string. Colors to be used to distinguish
different cluster.}
}
\value{
a list of the following items:
\itemize{
\item f: vector of f's.
\item delta: vector of delta's.
\item icenter: indices of cluster centers.
\item clusters: integer string recording cluster assignments. 
\item score: silouette of the clustering result.
\item h: bandwith h.
}
}
\description{
Uses f and delta (calculated from findRd()) to find clusters.
}
\details{
findCluster() plots f vs. delta plot, then allows user to select
centers of clusters and outliers.
}
\examples{
## Internal examples
}

