% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindClustersAuto.R
\name{FindClustersAuto}
\alias{FindClustersAuto}
\title{Automatically find cluster assignment that maximizes silouette.}
\usage{
FindClustersAuto(distm, f, delta, ac = 1, nclust = 2:10, f.cut = c(0.1,
  0.2, 0.3))
}
\arguments{
\item{distm}{distance matrix}

\item{f}{vector of local distance f(x)}

\item{delta}{vector of minimal distances to higher ground delta(x)}

\item{ac}{type of auto center picking from the f(x) vs. delta(x) plot. Only takes two values c(1, 2). If ac == 1 centers points with highest delta(x)'s to the right of vertical line at f(x) = f.cut. If ac == 2 centers are points above diagonal line.}

\item{nclust}{number of clusters. Can be a vector.}

\item{f.cut}{f(x) cuts to use when ac == 1. Must be values between 0 and 1. Can be a vector.}
}
\value{
list of four elements: clusters, centers, silhouette and nclust.
}
\description{
Automatically find cluster assignment that maximizes silouette.
}
\author{
Ethan Xu
}

