% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindClustersGivenCenters.R
\name{FindClustersGivenCenters}
\alias{FindClustersGivenCenters}
\title{Find cluster assignments given centers and distance matrix}
\usage{
FindClustersGivenCenters(distm, centers)
}
\arguments{
\item{distm}{distance matrix}

\item{centers}{vector of integers that gives the location of centers. Duplications will be silently dropped.}
}
\value{
vector of the same length as the the number of observations that gives cluster assignments with integers 1,2,...
}
\description{
Find cluster assignments given centers and distance matrix.
}
\author{
Ethan Xu
}

