% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindClustersManual.R
\name{FindClustersManual}
\alias{FindClustersManual}
\title{User-interactive routine to find clusters}
\usage{
FindClustersManual(distm, f, delta)
}
\arguments{
\item{distm}{distance matrix}

\item{f}{vector of local densities f(x). Same length of the number of observations}

\item{delta}{vector of distances to the closest high ground delta(x). Same length of the number of observations}
}
\value{
a list of the following items:
\itemize{
\item clusters: integer string recording cluster assignments. 
\item centers: indices of cluster centers.
\item score: silouette of the clustering result.
\item nclust: number of clusters.
}
}
\description{
Find clusters from user selections of centers on the
}
\details{
plots f(x) vs. delta(x) plot, then allows user to select
centers of clusters.
}
\examples{
## Internal examples
}

