% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDistm.R
\name{FindDistm}
\alias{FindDistm}
\title{Find distance matrix from data}
\usage{
FindDistm(x, normalize, method = "euclidean")
}
\arguments{
\item{x}{data}

\item{normalize}{boolean. normalize data before calculating distance?}

\item{method}{passed to 'dist()'}
}
\value{
distance matrix of class dist
}
\description{
Find distance matrix from data
}
\author{
Ethan Xu
}

