% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindFD.R
\name{FindFD}
\alias{FindFD}
\title{Find f and delta from distance matrix.}
\usage{
FindFD(distm, h, fdelta)
}
\arguments{
\item{distm}{distance matrix of class 'dist'}

\item{h}{bandwidth}

\item{fdelta}{String. One of [unorm, mnorm, weighted, count]}
}
\value{
list of f and delta. Each is a vector.
}
\description{
Find f(x) and delta(x) from distm, given h.
}
\details{
Find f(x) and delta(x).
}
\author{
Ethan Xu
}

