% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindH.R
\name{FindH}
\alias{FindH}
\title{Find bandwidth h}
\usage{
FindH(p, n, htype)
}
\arguments{
\item{p}{number of variables}

\item{n}{number of observations}

\item{htype}{methods to calculate h. One of the following (case insensitive) c("amise", "rot")}
}
\value{
bandwidth h.
}
\description{
Find bandwidth h
}
\author{
Ethan Xu
}

