% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROT.R
\name{ROT}
\alias{ROT}
\title{Calculate ROT bandwidth}
\usage{
ROT(x, y = NULL)
}
\arguments{
\item{x}{the number of variables (if y is missing), or a data frame or a matrix (if y is not missing).}

\item{y}{the number of observations. If y is missing, x should be the data matrix.}
}
\value{
ROT bandwidth.
}
\description{
Calculate the ROT bandwidth either from a data frame, or from p and n.
}
\details{
IMPORTANT NOTE: The standard deviation of each variable is omitted in this formula.
}

