% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.adpclust.R
\name{summary.adpclust}
\alias{summary.adpclust}
\title{Summary of adpclust}
\usage{
\method{summary}{adpclust}(object, ...)
}
\arguments{
\item{object}{object with class "adpclust". Return of adpclust()}

\item{...}{other arguments. NOT used in theis version.}
}
\description{
Summarizes results from the adpclust function from the ADPclust package.
}

