% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMISE.R
\name{AMISE}
\alias{AMISE}
\title{AMISE bandwidth}
\usage{
AMISE(x, y = NULL)
}
\arguments{
\item{x}{the number of variables (if y is missing), or a data frame or a matrix (if y is not missing).}

\item{y}{the number of observations. If y is missing, x should be the data matrix.}
}
\value{
AMISE bandwidth.
}
\description{
Calculate the AMISE bandwidth from either a data frame, or from p and n.
}
\details{
IMPORTANT NOTE: The standard deviation of each variable is omitted in the formula here. For 
computational consideration each dimension of the data is standardized before applying
the bandwidth. Two approaches are mathematically equivalent.
}

