% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust10.R
\docType{data}
\name{clust10}
\alias{clust10}
\title{1000 5-dim points scattered in 10 clusters}
\format{data frame}
\description{
1000 5-dim points in 10 clusters. Generated from the genRandomClust() function of the "clusterGeneration" package with separation value 0.2.
}
\keyword{datasets}

