% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clust3.R
\docType{data}
\name{clust3}
\alias{clust3}
\title{90 2-dim points scattered in 3 clusters}
\format{data frame}
\description{
90 2-dim points scattered in 3 clusters; Randomly generated from normal distributions
}
\keyword{datasets}

