% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCentersAutoD.R
\name{FindCentersAutoD}
\alias{FindCentersAutoD}
\title{Automatically finds centers with diagonal f(x) vs delta(x) thresholds}
\usage{
FindCentersAutoD(f, delta, nclust)
}
\arguments{
\item{f}{vector of local distance.}

\item{delta}{vector of minimal distances to higher ground.}

\item{nclust}{number of clusters. Can be a single integer or a vector of integers. Duplicates are silently removed.}
}
\value{
a list of vectors. Each vector gives the locations of centers.
}
\description{
Automatically finds centers with diagonal f(x) vs delta(x) thresholds. This is used in adpclust() with ac = 2. It finds points that are above and farthest from the diagonal line in the f vs. delta plots, and label them to be centers.
}
\author{
Ethan Xu
}

