% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindClustersGivenCenters.R
\name{FindClustersGivenCenters}
\alias{FindClustersGivenCenters}
\title{Find cluster assignments given centers and distance matrix}
\usage{
FindClustersGivenCenters(distm, centers)
}
\arguments{
\item{distm}{distance matrix}

\item{centers}{vector of integers that gives the indices of centers. Duplications will be silently dropped.}
}
\value{
Cluster assignments. A vector of the same length as the number of observations.
}
\description{
Find cluster assignments from given centers and distance matrix. Each point is assigned to the center that has the shortest Euclidean distance.
}

