% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.adpclust.R
\name{summary.adpclust}
\alias{summary.adpclust}
\title{Summary of adpclust}
\usage{
\method{summary}{adpclust}(object, ...)
}
\arguments{
\item{object}{object of class "adpclust" that is returned from adpclust().}

\item{...}{other arguments. NOT used.}
}
\description{
Summarizes the result from the adpclust() function.
}

